requirements.txt
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
six==1.17.0
psycopg2-binary
